+++ lib/blocks/has_execute_command.rb
@@ -5,6 +5,11 @@ module Bake
     module HasExecuteCommand
 
       def executeCommand(commandLine, ignoreStr=nil, exitCodeArray = [0], echo = "on")
+        if Bake.options.dry
+          puts commandLine
+          return true
+        end
+
         puts commandLine if (Bake.options.verbose >= 1 && echo != "off") || Bake.options.verbose >= 3
         puts "(executed in '#{@projectDir}')" if Bake.options.verbose >= 3
         cmd_result = false